requirements.txt
black==22.1.0
matplotlib==3.2.2
numpy==1.22.2
opencv_python==4.5.5.62
PyYAML==6.0
torch==1.10.2
torchvision==0.11.3
